+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ####05.07.2016
 --------------
+* [ADDED] New Console Command: /listplayers
+* [ADDED] Helperfunction isConsole(element)
 * [FIXED] #51 Some Fixes for Pizzajob
 
 ####04.07.2016